requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0

# HTTP clients
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# Message broker & Cache
kafka-python==2.0.2
redis==5.0.1

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Technical analysis
ta==0.11.0

# AI/ML
torch==2.1.1
transformers==4.35.2

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
